+++ libecc/source/common/soft-fp/double.h
@@ -60,7 +60,7 @@
 #define _FP_OVERFLOW_D		\
 	((_FP_W_TYPE)1 << _FP_WFRACBITS_D % _FP_W_TYPE_SIZE)
 
-typedef float DFtype __attribute__((mode(DF)));
+typedef double DFtype __attribute__((mode(DF)));
 
 #if _FP_W_TYPE_SIZE < 64
